move searchbar  to another package
