+++ app/src/main/kotlin/com/wire/android/ui/common/topappbar/search/AppTopBarWithSearchBarLayout.kt
@@ -1,4 +1,4 @@
-package com.wire.android.ui.common.topappbar
+package com.wire.android.ui.common.topappbar.search
 
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.AnimatedVisibility
@@ -23,12 +23,7 @@ import androidx.compose.material3.LocalTextStyle
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.res.painterResource
@@ -42,9 +37,6 @@ import com.wire.android.R
 import com.wire.android.ui.common.SearchBarInput
 import com.wire.android.ui.common.dimensions
 import com.wire.android.ui.theme.wireColorScheme
-import kotlinx.coroutines.flow.distinctUntilChanged
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.scan
 
 
 /**
@@ -233,41 +225,3 @@ private fun textStyleAlignment(isTopBarVisible: Boolean): TextStyle {
     )
 }
 
-@Composable
-private fun rememberSearchbarState(scrollPosition: Int): SearchBarState {
-    val searchBarState = remember {
-        SearchBarState()
-    }
-
-    LaunchedEffect(scrollPosition) {
-        snapshotFlow { scrollPosition }
-            .scan(0 to 0) { prevPair, newScrollIndex ->
-                if (prevPair.second == newScrollIndex || newScrollIndex == prevPair.second + 1) prevPair
-                else prevPair.second to newScrollIndex
-            }
-            .map { (prevScrollIndex, newScrollIndex) ->
-                newScrollIndex > prevScrollIndex + 1
-            }
-            .distinctUntilChanged().collect { shouldCollapse ->
-                searchBarState.isSearchBarCollapsed = shouldCollapse
-            }
-    }
-
-    return searchBarState
-}
-
-class SearchBarState {
-
-    var isSearchBarCollapsed by mutableStateOf(false)
-
-    var isTopBarVisible by mutableStateOf(true)
-        private set
-
-    fun hideTopBar() {
-        isTopBarVisible = false
-    }
-
-    fun showTopBar() {
-        isTopBarVisible = true
-    }
-}
